+++ examples/owned.rs
@@ -1,16 +1,15 @@
-//! Run using `cargo run --example simple --target x86_64-pc-windows-msvc`
+//! Run using `cargo run --example simple --target x86_64-pc-windows-msvc` --features=simulator
 //!
 //! Navigate using up/down arrows, interact using the Enter key
 
 use embedded_graphics::{pixelcolor::BinaryColor, prelude::Size, Drawable};
 use embedded_graphics_simulator::{
-    sdl2::Keycode, BinaryColorTheme, OutputSettingsBuilder, SimulatorDisplay, SimulatorEvent,
-    Window,
+    BinaryColorTheme, OutputSettingsBuilder, SimulatorDisplay, SimulatorEvent, Window,
 };
 use embedded_menu::{
-    interaction::InteractionType,
+    interaction::simulator::Simulator,
     items::{select::SelectValue, NavigationItem, Select},
-    Menu,
+    Menu, MenuStyle,
 };
 
 #[derive(Copy, Clone, PartialEq)]
@@ -45,16 +44,13 @@ fn main() -> Result<(), core::convert::Infallible> {
         .map(|i| Select::new(i, false))
         .collect();
 
-    let mut menu = Menu::new(format!("Items: {}", items.len()))
-        .add_item(
-            NavigationItem::new("Foo", ())
-                .with_marker(">")
-                .with_detail_text(
-                    "Lorem ipsum dolor sit amet, in per offendit assueverit adversarium, no sed clita adipisci nominati.",
-                ),
-        )
-        .add_items(items)
-        .build();
+    let mut menu = Menu::with_style(
+        format!("Items: {}", items.len()),
+        MenuStyle::new(BinaryColor::On).with_input_adapter(Simulator::default()),
+    )
+    .add_item(NavigationItem::new("Foo", ()).with_marker(">"))
+    .add_items(items)
+    .build();
 
     let output_settings = OutputSettingsBuilder::new()
         .theme(BinaryColorTheme::OledBlue)
@@ -68,20 +64,12 @@ fn main() -> Result<(), core::convert::Infallible> {
         window.update(&display);
 
         for event in window.events() {
+            menu.interact(event);
+
             match event {
-                SimulatorEvent::KeyDown {
-                    keycode,
-                    repeat: false,
-                    ..
-                } => match keycode {
-                    Keycode::Return => menu.interact(InteractionType::Select),
-                    Keycode::Up => menu.interact(InteractionType::Previous),
-                    Keycode::Down => menu.interact(InteractionType::Next),
-                    _ => None,
-                },
                 SimulatorEvent::Quit => break 'running,
-                _ => None,
-            };
+                _ => continue,
+            }
         }
     }